pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.0",
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "asyncio>=3.4.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "telegram>=0.0.1",
]
